prosignum/requirements.txt
Django==5.2.7
django-unfold==0.69.0
qrcode==8.2
pillow==12.0.0
python-decouple==3.8
psycopg2-binary==2.9.10
requests==2.32.3
